requirements.txt
streamlit
pandas
openpyxl
openai
google-generativeai
plotly

